+++ CHANGELOG.md
@@ -18,6 +18,7 @@ v1.6
 - Fix FLEXIAPI-252 Update the hCaptcha Laravel library, use file instead of cookies to store the session to prevent empty errors bags
 - Fix FLEXIAPI-254 Allow no data on POST requests to not trigger the ValidateJSON middleware
 - Fix FLEXIAPI-262 Bypass the JWT auth if we have an API Key
+- Fix FLEXIAPI-264 Add -k|api_key_ip parameter to accounts:create-admin-account to set/clear the related API Key restriction
 
 v1.5
 ---